requirements.txt
boto3==1.34.20
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
python-dotenv==1.0.1
pytest==7.4.4
pytest-mock==3.12.0
pytest-cov==4.1.0
moto==5.0.0
